+++ connectome_tools/s2f_recipe/estimate_bouton_reduction.py
@@ -26,7 +26,7 @@ class Executor(BaseExecutor):
     # while the function `sample_bouton_density` will make use of subprocesses
     is_parallel = False
 
-    def prepare(self, circuit, bio_data, sample=None):
+    def prepare(self, circuit, bio_data, sample=None, neurite_type=None):
         """Yield tasks that should be executed.
 
         Args:
@@ -35,15 +35,19 @@ class Executor(BaseExecutor):
                 or name of the .tsv file containing bouton density data (str).
             sample: sample configuration (dict)
                 or name of the .tsv file containing bouton density data (str).
+            neurite_type: name of the section that will be parsed (str)
+                if it is None, the default section will be "axon".
 
         Yields:
             (Task) task to be executed.
         """
         # pylint: disable=arguments-differ
-        yield Task(_execute, circuit, bio_data, sample, n_jobs=self.jobs, task_group=__name__)
+        yield Task(
+            _execute, circuit, bio_data, sample, neurite_type, n_jobs=self.jobs, task_group=__name__
+        )
 
 
-def _execute(circuit, bio_data, sample, n_jobs):
+def _execute(circuit, bio_data, sample, neurite_type, n_jobs):
     if isinstance(bio_data, float):
         ref_value = bio_data
     else:
@@ -59,6 +63,7 @@ def _execute(circuit, bio_data, sample, n_jobs):
         values = sample_bouton_density(
             circuit=circuit,
             n=sample.get("size", 100),
+            neurite_type=neurite_type,
             group=sample.get("target", None),
             mask=sample.get("mask", None),
             synapses_per_bouton=sample.get("assume_syns_bouton", 1.0),